// frontend/src/models/groupsModelType.ts
export interface groupsModelType extends Logic {
// ...
    values: {
        groupTypes: Array<GroupType>
        groupTypesLoading: boolean
        currentTeamId: number | null
        groupsEnabled: boolean
        groupsAccessStatus: GroupsAccessStatus
        showGroupsOptions: boolean
        groupsTaxonomicTypes: TaxonomicFilterGroupType[]
        groupNamesTaxonomicTypes: TaxonomicFilterGroupType[]
        aggregationLabel: (
            groupTypeIndex: number | null | undefined,
            deferToUserWording?: boolean
        ) => { singular: string; plural: string }
    }
// ...
}